refactor(stats): split 853-line _basic.py into core + composite mixins
Extract the leaf statistics (moments, averages, volatility, VaR/CVaR,
ulcer, win_rate) and shared static helpers into _BasicCoreMixin in
_stats/_basic_core.py. _BasicStatsMixin now subclasses it and keeps the
composite ratios, streak/outlier, and autocorrelation metrics that build
on the core. Dependency direction is one-way (composite → core), so the
inheritance is sound and the public mixin name is unchanged.
